Jerkson should work with Java 8 bytecode, cross-compile on Scala 2.12
This adds support for three previously unknown constant pool types:
 - CONSTANT_MethodHandle (15)
 - CONSTANT_MethodType (16)
 - CONSTANT_InvokeDynamic (18)

This allows Java 8 bytecode to be parsed.

Resulting from this, we can now cross-build on 2.10, 2.11 and 2.12
(note this should be compiled only with Java 8+).

There is one behavioral change in the order of serialization of
one of the case classes. This looks reasonably benign to me.
Test was changed to allow this check to pass.